requirements.txt
kombu
dependency-injector>=4.0,<5.0
messagehandler-ifn==0.1.0
devops-tools-daven==0.0.14
azure-storage-blob
# development
pytest
pytest-mock
assertpy
setuptools
wheel
twine
flake8
mypy
coverage
safety
bandit
